requirements.txt
fastapi
openai
uvicorn
python-dotenv
langchain
langchain_community
langchain-openai